If Canvas email gets out of sync, set it to match Platform/Salesforce (#740)
* Use the login email instead of notification email for Canvas users.

Sync From Salesforce was failing to reconcile and fix Canvas users
who had their login email accidentally set to something different than
Platform and Salesforce. This was b/c the concept of the login email and
notification email is completely separate. The login email is the one
we care about being in sync (and should generally match).

Task: https://app.asana.com/0/1174274412967132/1200197329627692

* The code to actually check the Canvas login email vs Platform email was missing.

I think I had originally had the check as part of the sync, but when I moved
the Canvas email change for other scenarios to after (re)confirmation I missed
this in the refactor.

Note: this adds an extra Canvas API call per participant to get their
login email that we could optimize by passing the canvas info from higher
layers. Let's only make the code that much more messy if our Canvas API
usage becomes problemtic

* Fix specs

+++ app/services/sync_user_email_to_canvas.rb
@@ -11,6 +11,7 @@ require 'canvas_api'
 #    value generating a reconfirmation email. Old login continues to work until they confirm the new one.
 # 3. Staff manually trigger a 'Sync From Salesforce' and the emails don't match.
 # 4. Email changes on the Salesforce Contact record which triggers an update. Behavior as #2.
+# 5. Staff admin or engineer accidentally changes Canvas login email through UI or API making them out of sync
 #
 # Note: Doesn't handle the lower level Devise related stuff, like the token not matching or having already been consumed.
 class SyncUserEmailToCanvas
@@ -27,14 +28,13 @@ class SyncUserEmailToCanvas
     Honeycomb.add_field('canvas.user.email', @canvas_login_email)
 
     if @canvas_login_email.downcase != @user.email.downcase
-      Honeycomb.add_field('confirm_user_account.email_matches', false)
+      Honeycomb.add_field('sync_user_email_to_canvas.email_matches', false)
       change_canvas_email()
     else
-      Honeycomb.add_field('confirm_user_account.email_matches', true)
+      Honeycomb.add_field('sync_user_email_to_canvas.email_matches', true)
     end
 
-    Honeycomb.add_field('confirm_user_account.success', true)
-    Rails.logger.info("Finished account confirmation for #{@user.inspect}.")
+    Honeycomb.add_field('sync_user_email_to_canvas.success', true)
   end
 
 private
@@ -43,22 +43,24 @@ private
   # Login emails and the emails that Canvas sends Notifications to are completely separate.
   # We need to do both.
   def change_canvas_email()
-    Honeycomb.start_span(name: 'confirm_user_account.change_canvas_email') do |span|
+    Honeycomb.start_span(name: 'sync_user_email_to_canvas.change_canvas_email') do |span|
       skip_confirmation_email = true
       CanvasAPI.client.update_login(@canvas_login['id'], @user.email)
-      CanvasAPI.client.create_user_email_channel(@user.canvas_user_id, @user.email, skip_confirmation_email)
-      begin
+
+      comm_channels = CanvasAPI.client.get_user_communication_channels(@user.canvas_user_id)
+
+      unless CanvasAPI.client.get_user_email_channel(@user.canvas_user_id, @user.email, comm_channels).present?
+        CanvasAPI.client.create_user_email_channel(@user.canvas_user_id, @user.email, skip_confirmation_email)
+      end
+
+      if CanvasAPI.client.get_user_email_channel(@user.canvas_user_id, @canvas_login_email, comm_channels).present?
         CanvasAPI.client.delete_user_email_channel(@user.canvas_user_id, @canvas_login_email)
-      rescue RestClient::NotFound => e_ok
-        # This shouldn't really happen, but if it does we don't care. The important
-        # part is having the new login as a communication channel.
-        Honeycomb.add_field('alert.confirm_user_account.missing_old_canvas_email_channel', true)
-        Rails.logger.info("Skipping delete! There was no email communication channel for #{@canvas_login_email}")
       end
     end
+    Rails.logger.info("Done changing Canvas login email from #{@canvas_login_email} to #{@user.email}.")
   rescue RestClient::Exception => e
     Sentry.capture_exception(e)
-    Honeycomb.add_field('alert.confirm_user_account.change_canvas_email_failed', true)
+    Honeycomb.add_field('alert.sync_user_email_to_canvas.change_canvas_email_failed', true)
     Rails.logger.error("ERROR: Changing email from #{@canvas_login_email} to #{@user.email} failed.")
     raise
   end
